Adding a custom script loader
Theres a ton of off-the-shelf libraries I could use to load a few
scripts. This is me just dabbling in building out a vanilla script
loader.

In additional to that, the app will always point to the latest
version of `dom-to-canvas.js`. No more mantaining a version in
the master and demo branches. :]